feat(#172): move mocks to test/mock
