+++ .github/workflows/go.yaml
@@ -41,13 +41,13 @@ jobs:
       run: go build
 
     - name: Test
-      run: go test -coverpkg='./internal/...,./web/...' -v './...'
+      run: go test -coverpkg='./internal/...' -v './...'
 
     - name: Vet
       run:  go vet ./...
 
     - name: Run coverage
-      run: go test -race -coverpkg='./internal/...,./web/...' -coverprofile=coverage.out -covermode=atomic ./...
+      run: go test -race -coverpkg='./internal/...' -coverprofile=coverage.out -covermode=atomic ./...
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
