+++ internal/repository/bitbucket/repository_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"github.com/Interhyp/metadata-service/internal/acorn/errors/httperror"
 	"github.com/Interhyp/metadata-service/internal/acorn/repository"
-	"github.com/Interhyp/metadata-service/test/acceptance/bbclientmock"
-	"github.com/Interhyp/metadata-service/test/acceptance/vaultmock"
+	"github.com/Interhyp/metadata-service/test/mock/bbclientmock"
+	"github.com/Interhyp/metadata-service/test/mock/vaultmock"
 	"github.com/StephanHCB/go-backend-service-common/repository/logging"
 	"github.com/stretchr/testify/require"
 	"testing"
