+++ test/acceptance/util_setup_test.go
@@ -11,11 +11,11 @@ import (
 	"github.com/Interhyp/metadata-service/internal/service/updater"
 	"github.com/Interhyp/metadata-service/internal/web/app"
 	"github.com/Interhyp/metadata-service/internal/web/server"
-	"github.com/Interhyp/metadata-service/test/acceptance/bitbucketmock"
-	"github.com/Interhyp/metadata-service/test/acceptance/idpmock"
-	"github.com/Interhyp/metadata-service/test/acceptance/kafkamock"
-	"github.com/Interhyp/metadata-service/test/acceptance/metadatamock"
-	"github.com/Interhyp/metadata-service/test/acceptance/vaultmock"
+	"github.com/Interhyp/metadata-service/test/mock/bitbucketmock"
+	"github.com/Interhyp/metadata-service/test/mock/idpmock"
+	"github.com/Interhyp/metadata-service/test/mock/kafkamock"
+	"github.com/Interhyp/metadata-service/test/mock/metadatamock"
+	"github.com/Interhyp/metadata-service/test/mock/vaultmock"
 	auacorn "github.com/StephanHCB/go-autumn-acorn-registry"
 	auconfigenv "github.com/StephanHCB/go-autumn-config-env"
 	librepo "github.com/StephanHCB/go-backend-service-common/acorns/repository"
