+++ test/mock/vaultmock/vault.go
